apps/narrative-ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.7.2
anthropicsdk==0.8.1
langchain==0.1.0
langchain-openai==0.0.2
aiohttp==3.9.1
python-multipart==0.0.6
